requirements.txt
black
pygments>=2.7.4
markdown>=3.3
matplotlib
jinja2
requests
mdx_truly_sane_lists
PyYAML
pandas
google-pasta
jupyter
pydot
boto3
tensorflow==2.19.1
keras-cv
keras-nlp
keras-tuner
tf-keras==2.19.0
keras-hub
keras-rs
pymdown-extensions==10.20
tensorflow-datasets

